Refactor auth reducers and remove redux-localstorage dependency
