+++ package.json
@@ -40,7 +40,7 @@
     "extract-text-webpack-plugin": "^2.1.0",
     "file-loader": "^0.11.1",
     "flux-standard-action": "^1.2.0",
-    "jest": "^20.0.3",
+    "jest": "^20.0.4",
     "json-loader": "0.5.4",
     "nock": "^9.0.9",
     "nodemon": "^1.11.0",
@@ -102,7 +102,6 @@
     "react-widgets": "^3.4.6",
     "redux": "3.6.0",
     "redux-actions": "^2.0.2",
-    "redux-localstorage": "^0.4.1",
     "redux-saga": "^0.15.3",
     "reselect": "^3.0.0",
     "serialize-javascript": "1.3.0"
