+++ src/containers/Logout/LogoutFederated.js
@@ -8,7 +8,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import * as actions from '../App/sessionActions';
+import * as actions from '../../modules/session/sessionActions';
 
 export class LogoutFederated extends React.Component {
   componentWillMount() {
