+++ src/containers/Logout/LogoutSession.js
@@ -9,7 +9,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import * as actions from '../App/sessionActions';
+import * as actions from '../../modules/session/sessionActions';
 
 export class LogoutSession extends React.Component {
   componentWillMount() {
