+++ src/sagas.js
@@ -9,7 +9,7 @@ import { spawn, all } from 'redux-saga/effects';
 import searchSagas from './containers/SearchPage/searchSagas';
 import articleSagas from './containers/TopicArticlePage/articleSagas';
 import tagSagas from './containers/Tag/tagSagas';
-import sessionSagas from './containers/App/sessionSagas';
+import sessionSagas from './modules/session/sessionSagas';
 import imageSearchSagas from './containers/ImageSearch/imageSearchSagas';
 
 export default function* root() {
